examples/requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
aiohttp==3.8.5
requests==2.31.0
soundfile==0.12.1